Set new branch with view to rebuild de-coupled version
Previous attempt had too much stateful data being passed around.
For example, the four global flags that various functions used to
keep track of the current calculation "stage". This was becoming
unwieldy from the standpoints of debugging and data hygiene.

View is to rewrite with the classes and functions to be more
loosely coupled, and less interdependent.

This separate branch takes it back to the base rendering logic,
before handling the input and processing, to approach again with
modularity in mind.